[TASK] [AB#130750] Add support for color scheme.
Note, this implementation is ugly because I didn't consider the extent of the need to update folders after creation, therefore folder updates are not atomic, which I don't believe is acccetable in production.
This is pretty easily fixed by extending the tagging service with a multi-update capability, allowing me to run the updates atomically.

That change will be backwards compatible though, so you can just start using this now.